style: 1.新增icon圖片  fix:調整DataBase架構   feat:新增記帳的類別區塊

+++ app/src/main/java/com/example/bookkeeping/model/keep_class/KeepClassDao.java
@@ -23,7 +23,7 @@ public interface KeepClassDao {
     void deleate(KeepClass keepClass);
 
     @Query("SELECT * FROM KeepClass WHERE kc_num LIKE :kc_num")
-    KeepClass findByID(Long kc_num);
+    KeepClass findByID(int kc_num);
 
     @Query("SELECT * FROM KeepClass")
     List<KeepClass> getAll();
